api/requirements/base.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
anyio==4.3.0
appnope==0.1.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
certifi==2023.5.7
cfgv==3.3.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.0.0
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-filter==23.2
django-graphql-jwt==0.3.4
django-import-export==3.3.7
django-mailersend==0.2.1
django-sms==0.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.12.2
frozenlist==1.3.3
graphene==3.2.2
graphene-django==3.1.1
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
httptools==0.6.1
identify==2.5.24
idna==3.4
ipython==7.25.0
jedi==0.19.1
jmespath==1.0.1
lxml==5.2.2
mailersend==0.5.1
MarkupPy==1.14
matplotlib-inline==0.1.7
multidict==6.0.4
nodeenv==1.8.0
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pip==24.0
platformdirs==3.8.0
pre-commit==2.21.0
promise==2.3
prompt-toolkit==3.0.43
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.17.2
PyJWT==2.7.0
pyotp==2.8.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
s3transfer==0.10.1
setuptools==67.8.0
six==1.16.0
sms==0.4
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
traitlets==5.14.3
twilio==8.3.0
typing_extensions==4.6.3
tzdata==2024.1
urllib3==1.26.16
virtualenv==20.23.1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2